Update 시소 짝꿍.cpp

+++ 프로그래머스/unrated/152996. 시소 짝꿍/시소 짝꿍.cpp
@@ -9,15 +9,15 @@ const int MAX = 1001;
     
 long long solution(vector<int> weights) {
     long long answer = 0;
-    long long cnt[MAX];
+    int cnt[MAX];
     fill_n(cnt, MAX, 0);
     for(int w : weights){
         cnt[w] += 1;
     }
     
     for(int w = 100; w < MAX; w++){
-        long long count = 0;
-        long long base = 0;
+        int count = 0;
+        int base = 0;
         //2:3
         if(w%2 == 0){
             base = w/2*3;
@@ -31,8 +31,8 @@ long long solution(vector<int> weights) {
         //2:4 = 1:2
         base = w*2;
         if(base < MAX) count += cnt[base];
-        answer += (count * cnt[w]);
-        answer += (cnt[w] * (cnt[w]-1)) /2;
+        answer += ((long long)count * cnt[w]);
+        answer += ((long long)cnt[w] * (cnt[w]-1)) /2;
     }
     return answer;
 }
